# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic restfull api for reddit-clone backend.",
        "title": "Reddit-Clone API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
        "/login": {
            "post": {
                "description": "Login via login and password in reddit-clone app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login to your account",
                "operationId": "login-user",
                "parameters": [
                    {
                        "description": "User credentials for authentication",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.AuthUserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User authorized successfully",
                        "schema": {
                            "$ref": "#/definitions/jwt.Session"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Bad login or password",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Register in reddit-clone app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "operationId": "register-user",
                "parameters": [
                    {
                        "description": "User credentials for registration",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users.AuthUserInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "schema": {
                            "$ref": "#/definitions/jwt.Session"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "422": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/errs.ComplexErrArr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.ComplexErr": {
            "description": "ComplexErr contains a more detailed description of the error, including the location and cause of the error",
            "type": "object",
            "properties": {
                "location": {
                    "description": "Any type"
                },
                "msg": {
                    "description": "Any type"
                },
                "param": {
                    "description": "Any type"
                },
                "value": {
                    "description": "Any type"
                }
            }
        },
        "errs.ComplexErrArr": {
            "description": "ComplexErrArr is an array of ComplexErr returned in case of a non-obvious error",
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errs.ComplexErr"
                    }
                }
            }
        },
        "errs.SimpleErr": {
            "description": "SimpleErr stores a brief description of an error",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Any type"
                }
            }
        },
        "jwt.Session": {
            "description": "Session stores the JWT token of the session",
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3MzUyMzU3ODAsImlhdCI6MTczNDYzMDk4MCwidXNlciI6eyJ1c2VybmFtZSI6InRlc3RfdXNlciIsImlkIjoiZDNkNzc1YmEtYTFlZS00MTEwLTkwOTktMTA0ZDVkYzFkYzQ2In19.I_3_yHlH1QUuKavtx8xVN_IRFMYXg3dYumzSrImA_NM"
                }
            }
        },
...
        "users.AuthUserInfo": {
            "description": "AuthUserInfo stores User credentials contained in the JWT Session token.",
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "example": "want_pizza"
                },
                "username": {
                    "type": "string",
                    "example": "Valery_Albertovich"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}